Develop

+++ app/src/main/java/com/example/achyp/weatherapp/pojo/Cloud.java
@@ -1,9 +1,6 @@
 package com.example.achyp.weatherapp.pojo;
 
-/**
- * Created by achyp on 2/18/17.
- */
 
 public class Cloud {
-    public int all;
+    public float all;
 }
