+++ app/src/main/java/com/example/achyp/weatherapp/pojo/Snow.java
@@ -1,9 +1,5 @@
 package com.example.achyp.weatherapp.pojo;
 
-/**
- * Created by achyp on 2/18/17.
- */
-
 public class Snow {
-    public int snow;
+    public float snow;
 }
